% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Preform  a very basic k-means clustering on a dataset.}
\usage{
k_means(dat, numclusters, PCA = F)
}
\arguments{
\item{dat}{numeric matrix to do kmeans to}

\item{numclusters}{number of clusters}

\item{PCA}{option to preform PCA on the data.}
}
\value{
A list containing cluster means, clustering vector, and total SSE.
}
\description{
Preform  a very basic k-means clustering on a dataset.
}
