% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Preform hierarchical clustering on a dataset.}
\usage{
hier_clust(x, method = "euclidean")
}
\arguments{
\item{x}{data for clustering
(such as a numeric vector or a data frame with all numeric columns).}

\item{method}{the distance measure to be used. This must be one of euclidean or manhattan.}
}
\value{
A vector with what was merged and in what order.
}
\description{
Preform hierarchical clustering on a dataset.
}
