% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Preform hierarchical clustering on a dataset.}
\usage{
hier_clust(dat, method = "euclidean")
}
\arguments{
\item{dat}{data for clustering (dataframe or matrix)}

\item{method}{the distance measure to be used. This must be one of euclidean or manhattan.}
}
\value{
A vector with what was merged and in what order.

A vector with the cluster assignments.
}
\description{
Preform hierarchical clustering on a dataset.
}
\examples{
iris2 <- iris \%>\% select(-Species)
hier_clust(iris2, method = "euclidean")
hier_clust(iris2, method = "manhattan")

}
